requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-ckeditor==5.6.1
django-cleanup==3.0.1
django-js-asset==1.1.0
django-mysql==2.4.1
gunicorn==19.9.0
mysqlclient==1.3.13
numpy==1.15.4
olefile==0.44
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.7
v==4.0.0
